+++ pkg/iam/nodes_template.go
@@ -14,6 +14,7 @@ const nodesReducedPermissionsTemplate = `{
         "ec2:DescribeInstances",
         "ec2:DescribeInstanceTypes",
         "ec2:DescribeNetworkInterfaces",
+		"ec2:DescribeRouteTables",
         "ec2:DescribeSecurityGroups",
         "ec2:DescribeSubnets",
         "ec2:DescribeTags",