+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.1] - 2025-09-05
+
 ### Changed
 
 - Add ec2:DescribeRouteTables permission to node template role for Cilium ENI mode.
@@ -354,7 +356,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement `AWSMachineTemplate` reconciler.
 - Implement `AWSMachinePool` reconciler.
 
-[Unreleased]: https://github.com/giantswarm/capa-iam-operator/compare/v1.4.0...HEAD
+[Unreleased]: https://github.com/giantswarm/capa-iam-operator/compare/v1.4.1...HEAD
+[1.4.1]: https://github.com/giantswarm/capa-iam-operator/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/giantswarm/capa-iam-operator/compare/v1.3.2...v1.4.0
 [1.3.2]: https://github.com/giantswarm/capa-iam-operator/compare/v1.3.1...v1.3.2
 [1.3.1]: https://github.com/giantswarm/capa-iam-operator/compare/v1.3.0...v1.3.1